Initial check-in
Getting this into TFS before all the work is lost. It's not done, and in
this state, it's not working either. By 'this state' I mean with the
asset service archetype generation. The tools code was compiling and
some of the tests were running when done as a stand-alone. The DMN weave
is mostly done. There may be some bits missing. The CMMN weave is only
partially done. The Signavio dictionary code was pulled out. The
extraction code is not yet done. The TrisotechWrapper is done for the
basics. The versions part needs to be discussed -- is it needed? and if
so, how to handle the way versions are returned.