Restructure
Restructured the files to remove superfluous modules to limit the number
of modules built in TFS. Moved ChainTest and supporting ChainConverter
to test under trisotech-metadata-extractor and removed trisotech-stage
module.
Builds in IntelliJ, but still complains in the pom for
metadata-extractor about kmdp-ext-terminology-content. Does NOT build
from the commandline for the same dependency.
There is added, commented-out code from Davide on some of the work that
will be done over the next few days.