More code review updates
Updated wrapper to return VersionIdentifier for getLatest. A number of
imports were updated to latest availble versions to enable compile.
afterWeave for CMMN updated to show the changes in the file now.
Fixed build issues. This resulted in updating some libraries, and this
resulted in DMN files no longer working for extraction because the
language cannot be properly detected. Debugged to an issue in Registry.
Reported to Davide. Disabled tests in the meantime until a fix is
available.