TFS #2035075 Refactor build
Refactored files and removed folders not needed to flatten structure for
TFS build. Also fixed artifactId name and made sure war file was being
created for service-impl.

TFS #2237517 Trisotech updates

Updated TrisotechWrapper to work with the new support for XML files to
be returned for Versions of Trisotech files.

TFS #2237517 plus Sonarlint

Ran sonarlint on the files to clean them up. Focused on TristechWrapper
and Weaver. There are still TODOS and some commented code that is there
to be handled at a later date. Also need to handle exceptions at a later
date -- do we have dedicated exceptions to use instead of generic?

Verified Weaver works with assetId out of the diagram and in the model.