TFS #2292744, #2286290, #2035075, #2035084
Updates to the previous commits to remove the token from the code base.
Now the token must be provided by an external file which is given on the
startup command.
To do this required setting up use of properties files, which
necessitated additional changes to the test and class files, in
particular to make them more 'Spring' friendly, using @Component and
@Autowired annotations.
A application-dev.properties file was added with the properties set for
the dev environment. The appication.properties file was modified to set
the active profile to dev (is this how we will do this?).
Config files were added for support of the test files.
A Static Initializer class was added to handle initialization of static
fields in another class as @Autowired will not work with static fields.
Doing it this way is not considered a best practice, but is sometimes
necessary. To discuss if there is a better, more desired way.