TFS #2369233
Updates made to support the ability to PUSH models back to Trisotech.
Expected use is with unpublished models, for use with the Dictionary.
Added test for published model to confirm it will work until use is
confirmed.

There is one bug with Trisotech waiting for a fix on that breaks the
model on update. There is another potential bug reported to Trisotech on
models being given state and version that are empty strings but return
in the API as published, but not in SPARQL. This causes a mismatch in
published models and will cause code to fail.

TFS #2369233

Fix call to upload to have correct version information.