TFS #2369233
Updates to the POST support to check for null state and version on both
the post and later on the filtering of published models. If a POST is
made with empty state and version parameters, those values will be set
to an empty string "" and will be returned to the APIs as having a
version and state even though they are empty. HOWEVER, the SPARQL query
will not return the state and version and therefore cannot be used to
find the assetId for the model.