TFS #2431057
Fix for bug issue. Code used in testing and initial diagramming needed
to pull out asset information from the diagram and put it in the root of
the document. With the new way to add the asset to the diagram, the data
is no longer in the diagram, but some of the test models still have it.
Should a model accidentally add it on the diagram in the future, keeping
this code would cause the workflow to fail due to duplicate information
in the JSON. Removing the code removes this issue.
Changed a couple of tests to not run until Trisotech fix is received.