TFS #2431057
Fix for issue with older models that had the asset on the diagram which
meant the assettype information was in the diagram section. Now
assettype information is in the root of the document and no longer needs
to be pulled out of the diagram. If an older diagram is used that still
has the diagram information as well as in the root of the document, it
will result in a duplication JSON error when retrieving the data from
the API. Removing the code that moved this from the diagram and placed
into the root fixes the issue.

Some tests were updated due to additional models being added to the test
repository.