TFS #2369233
Disabled hierarchySorter for now until can come up with a new solution.
Re-enabled tests for POST and confirmed POST is now working with
Trisotech fix.
HierarchySorter had to be disabled due to Trisotech fix including a
change to the triples such that the semantic links are now
bi-directional. This causes an infinite loop and eventually a
StackOverflow exception that breaks the starting of the application.