TFS #2563586
Some more updates:
Fixed SPARQL query to only return the assetId that is of the appropriate
type. For DMN, this is decision. For CMMN, this is activity. For CMMN,
it may not be necessary, but also does not appear to hurt.
Cleaned up some of the queryModels to not include select values that are
not actually generated in the query.
Fixed surrogate output for dependent artifact name.