AB #62158
Updates to add timestamp to the version of the artifacts in the
surrogate information.
Added the support to generate models from MEA for testing, etc. so will
be available when I am not.
Updated some test resource files with latest information.
This builds on Java 11.
Added a README
Added support for handling not latest version of an artifact and getting
the appropriate not latest versions of the dependencies.