AB #62158
Updates to add timestamp to the version of the artifacts in the
surrogate information.
Added the support to generate models from MEA for testing, etc. so
will be available when I am not.
Updated some test resource files with latest information.
This builds on Java 11.
Added a README
Added support for handling not latest version of an artifact and
getting the appropriate not latest versions of the dependencies.
For some reason a number of files were missed in the last push.
Some dependency versions have been updated.
Disabled for now the use of hierarchySorter as it is causing infinite
loops again, this time in both MEA and MEA-Test. More investigation
needed.