AB #341535
Updated Trisotech Wrapper (TTW) for v10 of the Trisotech. Most of the
updates were to the SPARQL queries. The other APIs only needed to be
modified where fileId was used as fileId is now the same URI as the
modelId.
All the models were updated making use of the copyOf and reuse that is now in
Trisotech.
The weaver was updated to handle the new trisotech tags.
The code is currently hardcoded to the test-mc endpoint and to make use
of a different token (test-trisotechToken). These are in place so we can
test and verify the models and the server before the switch on the
Trisotech side. At that time, the code will need to be updated to use
the existing endpoint and token.
Rebased with latest develop branch.